Extracted CommonBackendContext, common Ir and Symbols
